requirements.txt
python-dotenv
requests
numpy

